requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.4,<5.0
redis>=5.0,<8.0
